Add Matchers class
self.match is the primary public-facing method. It takes a query in the
form of a hash, where the top-level keys are the names of matcher
strategies and the values are the arguments for each matcher. The
corresponding method is called with the arguments. For
example,

  Matchers.match(node, assignment: { lvalue: :title })

would call

  Matchers.assignment(lvalue: :title)

Currently, a node matches only if all components of the query match.
This should be specifiable.